Enable ruff RUF046 rule and remove unnecessary int() casts (#170514)

+++ homeassistant/components/bluesound/media_player.py
@@ -689,7 +689,7 @@ class BluesoundPlayer(CoordinatorEntity[BluesoundCoordinator], MediaPlayerEntity
 
     async def async_set_volume_level(self, volume: float) -> None:
         """Send volume_up command to media player."""
-        volume = int(round(volume * 100))
+        volume = round(volume * 100)
         volume = min(100, volume)
         volume = max(0, volume)
